KVM: arm/arm64: vgic-new: Add basic code frame and add to build system
Adds the basic code frame in a new directory and implements the bare
necessities. This includes vgic_get_irq, which looks up an irq-struct
based on a VM pointer and an interrupt ID, and potentially a VCPU
pointer for SGIs and PPIs.
kvm_vgic_inject_irq and vgic_update_irq_pending are provided as
skeletons.
Locking is rather important here to avoid races but still remain
scalable, so we have a lock per IRQ, and we have a lock per vcpu's
ap_list.
To prevent any deadlock we enforce the following locking order rules:
- Always take the ap_list_lock before the vgic_irq lock
- When taking more than one ap_list_lock at the same time, always take
the lowest numbered VCPU's ap_list_lock first.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Eric Auger <eric.auger@linaro.org>
Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>
Signed-off-by: Andre Przywara <andre.przywara@arm.com>