KVM: arm/arm64: vgic-new: Export register access interface
Userland can access the emulated GIC to save and restore its state
for initialization or migration purposes.
Provide an interface to let userland access the register handlers
without using the kvm_io_bus interface, since this requires knowledge
of the absolute guest address, which this userland interface does not
provide.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>