NOT FOR UPSTREAM: pine64.dts: add memreserve
The existing Allwinner firmware loads various firmware blobs all over
the place, so it needs to protect them from the kernel.
Actually the firmware should patch the DT to match the actual
configuration, but this does not seem to happen, so have these values
here instead.
The need for this hack will go away once we have sane firmware.