arm64: dts: sun50i-a64: add sun50i-emac Ethernet driver
The sun50i-emac is an Ethernet MAC that supports 10/100/1000 Mbit
connections. It is very similar to the device found in the Allwinner
H3, but lacks the internal 100 Mbit PHY and its associated control
bits.
This adds the necessary bits to the Allwinner A64 SoC .dtsi, but keeps
it disabled at this level.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>